get.sylly.env {sylly}	R Documentation

Get sylly session settings

Description

The function get.sylly.env returns information on your session environment regarding the sylly package, e.g. whether a cache file should be used, if it was set before using set.sylly.env.

Usage

get.sylly.env(..., errorIfUnset = TRUE)

Arguments

...
Named parameters to get from the sylly environment. Valid arguments are:

lang

Logical, whether the set language should be returned.

hyph.cache.file

Logical, whether the set hyphenation cache file for hyphen should be returned.

hyph.max.token.length

Logical, whether the set maximum token length should be returned.

errorIfUnset
Logical, if TRUE and the desired property is not set at all, the function will fail with an error message.

Details

For the most part, get.sylly.env is a convenient wrapper for getOption.

Value

A character string or list, possibly including:

lang
The specified language

hyph.cache.file
The specified hyphenation cache file for hyphen

See Also

set.sylly.env

Examples

set.sylly.env(hyph.cache.file=file.path(tempdir(), "cache_file.RData"))
get.sylly.env(hyph.cache.file=TRUE)

[Package sylly version 0.1-6 Index]