install.sylly.lang {sylly}	R Documentation

Install language support packages

Description

This is a wrapper for install.packages, making it more convenient to install additional language support packages for sylly.

Usage

install.sylly.lang(
  lang,
  repos = "https://undocumeantit.github.io/repos/l10n/",
  ...
)

Arguments

lang
Character vector, one or more valid language identifiers (like en for English or de for German).

repos
The URL to additional repositories to query. You should probably leave this to the default, but if you would like to use a third party repository, you're free to do so. The value is temporarily appended to the repos currently returned by getOption("repos").

...
Additional options for install.packages.

Details

For a list of currently available language packages see available.sylly.lang. See set.hyph.support for more details on sylly's language support in general.

Value

Does not return any useful objects, just calls install.packages.

See Also

install.packages, available.sylly.lang

Examples

## Not run: 
# install support for German
install.sylly.lang("de")
# load the package
library("sylly.de")

## End(Not run)

[Package sylly version 0.1-6 Index]