kRp.hyph.pat,-class {sylly}	R Documentation

S4 Class kRp.hyph.pat

Description

This class is used for objects that are returned by read.hyph.pat.

Details

Since this package has been a part of the koRpus package before, you might run into old pattern files. You will know that this is the case if using them automatically tries to load the koRpus package. In these cases, you might want to strip the defunct reference to koRpus by calling the private function sylly:::koRpus2sylly which take the path to the old file as its first argument. Be aware that calling this function will overwrite the old file in-place, so you should make a backup first!

Slots

lang

A character string, naming the language that is assumed for the patterns in this object

pattern

A matrix with three colums:

orig:

The unchanged patgen patterns.

char:

Only the characters used for matching.

nums:

The hyphenation number code for the pattern.

Contructor function

Should you need to manually generate objects of this class (which should rarely be the case), the contructor function kRp_hyph_pat(...) can be used instead of new("kRp.hyph.pat", ...). Whenever possible, stick to read.hyph.pat.

[Package sylly version 0.1-6 Index]