kRp.hyphen,-class {sylly}	R Documentation

S4 Class kRp.hyphen

Description

This class is used for objects that are returned by hyphen.

Slots

lang

A character string, naming the language that is assumed for the analized text in this object

desc

Descriptive statistics of the analyzed text.

hyphen

A data.frame with two columns:

syll:

Number of recognized syllables

word:

The hyphenated word

Contructor function

Should you need to manually generate objects of this class (which should rarely be the case), the contructor function kRp_hyphen(...) can be used instead of new("kRp.hyphen", ...). Whenever possible, stick to hyphen.

[Package sylly version 0.1-6 Index]