describe {sylly}	R Documentation

Getter/setter methods for sylly objects

Description

These methods should be used to get or set values of hyphenated text objects generated by functions like hyphen().

Usage

describe(obj, ...)

## S4 method for signature 'kRp.hyphen'
describe(obj)

describe(obj, ...) <- value

## S4 replacement method for signature 'kRp.hyphen'
describe(obj, ...) <- value

language(obj)

## S4 method for signature 'kRp.hyphen'
language(obj)

language(obj) <- value

## S4 replacement method for signature 'kRp.hyphen'
language(obj) <- value

hyphenText(obj)

## S4 method for signature 'kRp.hyphen'
hyphenText(obj)

hyphenText(obj) <- value

## S4 replacement method for signature 'kRp.hyphen'
hyphenText(obj) <- value

## S4 method for signature 'kRp.hyphen'
x[i, j]

## S4 replacement method for signature 'kRp.hyphen'
x[i, j] <- value

## S4 method for signature 'kRp.hyphen'
x[[i]]

## S4 replacement method for signature 'kRp.hyphen'
x[[i]] <- value

Arguments

obj
An object of class kRp.hyphen.

...
Additional arguments as defined by respective methods.

value
A value to set.

x
An object of class kRp.hyphen.

i
Row index.

j
Column index.

Details

describe() returns the desc slot.

language() returns the lang slot.

hyphenText() returns the hyphen slot from objects of class kRp.hyphen.

[/[[ Can be used as a shortcut to index the results of hyphenText().

Examples

## Not run: 
hyphenText(hyphenated.txt)

## End(Not run)

[Package sylly version 0.1-6 Index]