sylly-package {sylly}	R Documentation

Hyphenation and Syllable Counting for Text Analysis

Description

Provides the hyphenation algorithm used for 'TeX'/'LaTeX' and similar software, as proposed by Liang (1983, <https://tug.org/docs/liang/>). Mainly contains the function hyphen() to be used for hyphenation/syllable counting of text objects. It was originally developed for and part of the 'koRpus' package, but later released as a separate package so it's lighter to have this particular functionality available for other packages. Support for various languages needs be added on-the-fly or by plugin packages (<https://undocumeantit.github.io/repos/>); this package does not include any language specific data. Due to some restrictions on CRAN, the full package sources are only available from the project homepage. To ask for help, report bugs, request features, or discuss the development of the package, please subscribe to the koRpus-dev mailing list (<http://korpusml.reaktanz.de>).

Details

The DESCRIPTION file:

Package: 	 sylly

Type: 	 Package

Version: 	 0.1-6

Date: 	 2020-09-19

Depends: 	 R (>= 3.0.0)

Encoding: 	 UTF-8

License: 	 GPL (>= 3)

LazyLoad: 	 yes

URL: 	 https://reaktanz.de/?c=hacking&s=sylly

Author(s)

NA

Maintainer: NA

See Also

Useful links:

https://reaktanz.de/?c=hacking&s=sylly

Report bugs at https://github.com/unDocUMeantIt/sylly/issues

[Package sylly version 0.1-6 Index]